test(src): move final inline cfg(test) modules to tests dir (closes #146)
