+++ src/executor/secret_transport.rs
@@ -97,48 +97,3 @@ impl Drop for SecretHandle {
         }
     }
 }
-
-#[cfg(test)]
-mod inline_tests {
-    use super::EphemeralSecretFile;
-    use std::panic;
-
-    #[test]
-    fn file_exists_after_write() {
-        let values = vec![("EXIST".to_string(), "val".to_string())];
-        let handle = EphemeralSecretFile::write(&values).expect("write");
-        assert!(handle.paths()[0].exists(), "file must exist");
-        drop(handle);
-    }
-
-    #[test]
-    fn file_removed_on_drop() {
-        let values = vec![("DROP".to_string(), "val".to_string())];
-        let path = {
-            let handle = EphemeralSecretFile::write(&values).expect("write");
-            handle.paths()[0].clone()
-        };
-        assert!(!path.exists(), "file must be dropped");
-    }
-
-    #[test]
-    fn file_removed_after_panic() {
-        let values = vec![("PANIC".to_string(), "val".to_string())];
-        let path = {
-            let handle = EphemeralSecretFile::write(&values).expect("write");
-            let p = handle.paths()[0].clone();
-            assert!(p.exists());
-            let result = panic::catch_unwind(panic::AssertUnwindSafe(|| {
-                let _h = handle;
-                panic!("simulated panic in test");
-            }));
-            assert!(result.is_err());
-            p
-        };
-        assert!(
-            !path.exists(),
-            "file must be removed after panic: {}",
-            path.display()
-        );
-    }
-}
