+++ src/scheduler/circuit.rs
@@ -489,305 +489,3 @@ impl CircuitStore {
         Ok(record)
     }
 }
-
-#[cfg(test)]
-mod inline_tests {
-    use super::{
-        AdmissionResult, CircuitConfig, CircuitScope, CircuitState, CircuitStore, Connection,
-    };
-    use crate::daemon::orchestration::FakeClock;
-
-    fn circuit_store() -> (CircuitStore, FakeClock) {
-        let conn = Connection::open_in_memory().expect("in-memory db");
-        conn.execute_batch(
-            "CREATE TABLE IF NOT EXISTS circuit_state (
-                scope TEXT NOT NULL,
-                scope_id TEXT NOT NULL,
-                state TEXT NOT NULL DEFAULT 'closed',
-                consecutive_failures INTEGER NOT NULL DEFAULT 0,
-                last_failure_at INTEGER,
-                opened_at INTEGER,
-                last_probe_at INTEGER,
-                PRIMARY KEY (scope, scope_id)
-            );",
-        )
-        .expect("create circuit_state table");
-        let config = CircuitConfig::test_defaults();
-        let store = CircuitStore::new(conn, config);
-        let clock = FakeClock::new();
-        (store, clock)
-    }
-
-    #[test]
-    fn circuit_state_round_trip() {
-        assert_eq!(CircuitState::Closed.as_str(), "closed");
-        assert_eq!(CircuitState::Open.as_str(), "open");
-        assert_eq!(CircuitState::HalfOpen.as_str(), "half_open");
-        assert_eq!(
-            "closed".parse::<CircuitState>().unwrap(),
-            CircuitState::Closed
-        );
-        assert_eq!("open".parse::<CircuitState>().unwrap(), CircuitState::Open);
-        assert_eq!(
-            "half_open".parse::<CircuitState>().unwrap(),
-            CircuitState::HalfOpen
-        );
-        assert!("invalid".parse::<CircuitState>().is_err());
-    }
-
-    #[test]
-    fn circuit_scope_round_trip() {
-        assert_eq!(CircuitScope::Provider.as_str(), "provider");
-        assert_eq!(CircuitScope::Repository.as_str(), "repository");
-        assert_eq!(
-            "provider".parse::<CircuitScope>().unwrap(),
-            CircuitScope::Provider
-        );
-        assert_eq!(
-            "repository".parse::<CircuitScope>().unwrap(),
-            CircuitScope::Repository
-        );
-        assert!("invalid".parse::<CircuitScope>().is_err());
-    }
-
-    #[test]
-    fn threshold_opens_circuit() {
-        // GIVEN a closed circuit for repository "owner/repo"
-        let (store, clock) = circuit_store();
-
-        // WHEN 3 consecutive failures occur
-        let r1 = store
-            .record_failure("repository", "owner/repo", &clock)
-            .unwrap();
-        assert_eq!(r1.state, CircuitState::Closed);
-        assert_eq!(r1.consecutive_failures, 1);
-
-        let r2 = store
-            .record_failure("repository", "owner/repo", &clock)
-            .unwrap();
-        assert_eq!(r2.state, CircuitState::Closed);
-        assert_eq!(r2.consecutive_failures, 2);
-
-        let r3 = store
-            .record_failure("repository", "owner/repo", &clock)
-            .unwrap();
-        // THEN the circuit opens
-        assert_eq!(r3.state, CircuitState::Open);
-        assert_eq!(r3.consecutive_failures, 3);
-        assert!(r3.opened_at.is_some());
-    }
-
-    #[test]
-    fn closed_circuit_admits() {
-        // GIVEN a closed circuit
-        let (store, clock) = circuit_store();
-
-        // WHEN try_admit is called
-        let result = store.try_admit("repository", "owner/repo", &clock).unwrap();
-
-        // THEN admission is granted
-        assert_eq!(result, AdmissionResult::NoCircuit);
-    }
-
-    #[test]
-    fn open_circuit_rejects_admission() {
-        // GIVEN an open circuit (3 failures)
-        let (store, clock) = circuit_store();
-        for _ in 0..3 {
-            let _ = store
-                .record_failure("repository", "owner/repo", &clock)
-                .unwrap();
-        }
-
-        // WHEN try_admit is called immediately
-        let result = store.try_admit("repository", "owner/repo", &clock).unwrap();
-
-        // THEN admission is rejected with CircuitOpen
-        match result {
-            AdmissionResult::CircuitOpen {
-                retry_after,
-                probe_in_flight,
-            } => {
-                assert!(retry_after > 0);
-                assert!(!probe_in_flight);
-            }
-            other => panic!("expected CircuitOpen, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn half_open_transition_after_open_interval() {
-        // GIVEN an open circuit
-        let (store, clock) = circuit_store();
-        for _ in 0..3 {
-            let _ = store
-                .record_failure("repository", "owner/repo", &clock)
-                .unwrap();
-        }
-
-        // WHEN the open interval elapses
-        clock.advance(1800);
-
-        // THEN try_admit transitions to HalfOpen and admits
-        let result = store.try_admit("repository", "owner/repo", &clock).unwrap();
-        assert_eq!(result, AdmissionResult::Admitted);
-
-        // Verify the circuit is now HalfOpen
-        let record = store
-            .get_record("repository", "owner/repo")
-            .unwrap()
-            .unwrap();
-        assert_eq!(record.state, CircuitState::HalfOpen);
-    }
-
-    #[test]
-    fn successful_probe_resets_circuit() {
-        // GIVEN an open circuit that transitions to HalfOpen
-        let (store, clock) = circuit_store();
-        for _ in 0..3 {
-            let _ = store
-                .record_failure("repository", "owner/repo", &clock)
-                .unwrap();
-        }
-        clock.advance(1800);
-        let _ = store.try_admit("repository", "owner/repo", &clock).unwrap();
-
-        // WHEN the probe succeeds
-        store
-            .record_probe_result("repository", "owner/repo", true, &clock)
-            .unwrap();
-
-        // THEN the circuit resets to Closed
-        let record = store
-            .get_record("repository", "owner/repo")
-            .unwrap()
-            .unwrap();
-        assert_eq!(record.state, CircuitState::Closed);
-        assert_eq!(record.consecutive_failures, 0);
-    }
-
-    #[test]
-    fn failed_probe_reopens_circuit() {
-        // GIVEN an open circuit that transitions to HalfOpen
-        let (store, clock) = circuit_store();
-        for _ in 0..3 {
-            let _ = store
-                .record_failure("repository", "owner/repo", &clock)
-                .unwrap();
-        }
-        clock.advance(1800);
-        let _ = store.try_admit("repository", "owner/repo", &clock).unwrap();
-
-        // WHEN the probe fails
-        store
-            .record_probe_result("repository", "owner/repo", false, &clock)
-            .unwrap();
-
-        // THEN the circuit returns to Open
-        let record = store
-            .get_record("repository", "owner/repo")
-            .unwrap()
-            .unwrap();
-        assert_eq!(record.state, CircuitState::Open);
-        assert!(record.opened_at.is_some());
-    }
-
-    #[test]
-    fn twentyfour_hour_age_triggers_escalation() {
-        // GIVEN an open circuit
-        let (store, clock) = circuit_store();
-        for _ in 0..3 {
-            let _ = store
-                .record_failure("repository", "owner/repo", &clock)
-                .unwrap();
-        }
-
-        // WHEN 24 hours elapse
-        clock.advance(86400);
-
-        // THEN try_admit returns MaxDegradedAgeExceeded
-        let result = store.try_admit("repository", "owner/repo", &clock).unwrap();
-        assert_eq!(result, AdmissionResult::MaxDegradedAgeExceeded);
-
-        // AND exhausted_to_needs_attention returns the entry
-        let exhausted = store.exhausted_to_needs_attention(&clock).unwrap();
-        assert_eq!(exhausted.len(), 1);
-        assert_eq!(exhausted[0].scope_id, "owner/repo");
-    }
-
-    #[test]
-    fn open_less_than_24h_does_not_escalate() {
-        // GIVEN an open circuit
-        let (store, clock) = circuit_store();
-        for _ in 0..3 {
-            let _ = store
-                .record_failure("repository", "owner/repo", &clock)
-                .unwrap();
-        }
-
-        // WHEN only 5 minutes elapse (less than open interval of 30 min)
-        clock.advance(300);
-
-        // THEN try_admit does NOT return MaxDegradedAgeExceeded
-        let result = store.try_admit("repository", "owner/repo", &clock).unwrap();
-        match result {
-            AdmissionResult::CircuitOpen { .. } => {} // expected
-            other => panic!("expected CircuitOpen, got {other:?}"),
-        }
-
-        // AND exhausted_to_needs_attention returns empty
-        let exhausted = store.exhausted_to_needs_attention(&clock).unwrap();
-        assert!(exhausted.is_empty());
-    }
-
-    #[test]
-    fn restart_preserves_state() {
-        // GIVEN an open circuit saved to persistent SQLite
-        let dir = std::env::temp_dir().join(format!("circuit-test-{}", std::process::id()));
-        let _ = std::fs::remove_dir_all(&dir);
-        std::fs::create_dir_all(&dir).unwrap();
-        let db_path = dir.join("circuit.db");
-
-        let config = CircuitConfig::test_defaults();
-        let clock = FakeClock::new();
-
-        {
-            let conn = Connection::open(&db_path).expect("open db");
-            conn.execute_batch(
-                "CREATE TABLE IF NOT EXISTS circuit_state (
-                    scope TEXT NOT NULL,
-                    scope_id TEXT NOT NULL,
-                    state TEXT NOT NULL DEFAULT 'closed',
-                    consecutive_failures INTEGER NOT NULL DEFAULT 0,
-                    last_failure_at INTEGER,
-                    opened_at INTEGER,
-                    last_probe_at INTEGER,
-                    PRIMARY KEY (scope, scope_id)
-                );",
-            )
-            .expect("create table");
-            let store = CircuitStore::new(conn, config.clone());
-            for _ in 0..3 {
-                let _ = store
-                    .record_failure("repository", "owner/repo", &clock)
-                    .unwrap();
-            }
-        }
-
-        // WHEN the connection is closed and a new one opens
-        {
-            let conn = Connection::open(&db_path).expect("re-open db");
-            let store = CircuitStore::new(conn, config);
-
-            // THEN the circuit state is preserved
-            let record = store
-                .get_record("repository", "owner/repo")
-                .unwrap()
-                .unwrap();
-            assert_eq!(record.state, CircuitState::Open);
-            assert_eq!(record.consecutive_failures, 3);
-        }
-
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-}
